+++ src/renderer/components/header/HeaderPriceSelector.stories.tsx
@@ -6,7 +6,7 @@ import { storiesOf } from '@storybook/react'
 import { PoolAsset, PricePoolAssets, PricePoolAsset } from '../../views/pools/types'
 import HeaderPriceSelector from './HeaderPriceSelector'
 
-const assets: PricePoolAssets = [PoolAsset.RUNE67C, PoolAsset.BTC, PoolAsset.ETH, PoolAsset.TUSDB]
+const assets: PricePoolAssets = [PoolAsset.RUNE67C, PoolAsset.BTC, PoolAsset.ETH, PoolAsset.BUSDB]
 
 storiesOf('Components/HeaderPriceSelector', module).add('desktop / mobile', () => {
   const changeHandler = (asset: PricePoolAsset) => console.log(`changed: ${asset}`)
@@ -16,7 +16,7 @@ storiesOf('Components/HeaderPriceSelector', module).add('desktop / mobile', () =
       assets={assets}
       isDesktopView={isDesktopView}
       changeHandler={changeHandler}
-      selectedAsset={PoolAsset.TUSDB}
+      selectedAsset={PoolAsset.BUSDB}
     />
   )
 })
