+++ src/renderer/components/header/util.ts
@@ -10,7 +10,7 @@ import { PoolAsset, PricePoolAsset } from '../../views/pools/types'
 export const toHeaderCurrencyLabel = (asset: PricePoolAsset): string => {
   let ticker = assetFromString(asset)?.ticker ?? 'unknown'
   // special case TUSDB
-  if (asset === PoolAsset.TUSDB) ticker = 'USD'
+  if (asset === PoolAsset.BUSDB) ticker = 'USD'
   return `${CURRENCY_SYMBOLS[asset]} ${ticker}`
 }
